move module onboarding screen
