+++ ui/src/main/java/com/jin/ui/onboarding/OnboardingScreen.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.onboarding.ui
+package com.jin.ui.onboarding
 
 import android.Manifest
 import androidx.compose.foundation.Image
@@ -14,8 +14,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.LocationOn
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.HorizontalDivider
@@ -26,7 +24,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
@@ -39,7 +36,7 @@ import androidx.compose.ui.unit.sp
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.PermissionStatus
 import com.google.accompanist.permissions.rememberPermissionState
-import com.jin.honey.R
+import com.jin.ui.R
 import com.jin.ui.theme.HoneyTheme
 import com.jin.ui.theme.OnboardingDescTextColor
 import com.jin.ui.theme.OnboardingIconBackgroundColor
